LayoutTests/imported/w3c/web-platform-tests/css/css-break/table/border-spacing-expected.txt

No red should be seen below.

1
2
3
4
5

FAIL Table with border spacing assert_equals: expected 110 but got 10
